Add files via upload
This is a fully functional release. I've implemented a random opponent builder with 15 different monsters types and a player leveling system. I havent written in any save functions, so unfortunatly you start with a new character everytime, but that will be changed soon. This is just the base of the project I have planned so watch for updates